+++ crates/spirv-std/src/image.rs
@@ -1,6 +1,8 @@
 //! Image types
 
 #[cfg(target_arch = "spirv")]
+use crate::vector::VectorTruncateInto;
+#[cfg(target_arch = "spirv")]
 use core::arch::asm;
 
 // Rustfmt formats long marker trait impls over multiple lines which makes them
@@ -97,13 +99,14 @@ pub type Cubemap = crate::Image!(cube, type=f32, sampled, __crate_root=crate);
 // HACK(eddyb) avoids "transparent newtype of `_anti_zst_padding`" misinterpretation.
 #[repr(C)]
 pub struct Image<
-    SampledType: SampleType<FORMAT>,
+    SampledType: SampleType<FORMAT, COMPONENTS>,
     const DIM: u32,          // Dimensionality,
     const DEPTH: u32,        // ImageDepth,
     const ARRAYED: u32,      // Arrayed,
     const MULTISAMPLED: u32, // Multisampled,
     const SAMPLED: u32,      // Sampled,
     const FORMAT: u32,       // ImageFormat,
+    const COMPONENTS: u32,   // NumberOfComponents,
 > {
     // HACK(eddyb) avoids the layout becoming ZST (and being elided in one way
     // or another, before `#[spirv(generic_image_type)]` can special-case it).
@@ -112,22 +115,36 @@ pub struct Image<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const FORMAT: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::Yes as u32 }, FORMAT>
+        const COMPONENTS: u32,
+    >
+    Image<
+        SampledType,
+        DIM,
+        DEPTH,
+        ARRAYED,
+        MULTISAMPLED,
+        { Sampled::Yes as u32 },
+        FORMAT,
+        COMPONENTS,
+    >
 {
     /// Fetch a single texel with a sampler set at compile time
     #[crate::macros::gpu_only]
     #[doc(alias = "OpImageFetch")]
-    pub fn fetch<I>(&self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED>) -> SampledType::Vec4
+    pub fn fetch<I>(
+        &self,
+        coordinate: impl ImageCoordinate<I, DIM, ARRAYED>,
+    ) -> SampledType::SampleResult
     where
         I: Integer,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
         unsafe {
             asm! {
                 "%image = OpLoad _ {this}",
@@ -139,18 +156,29 @@ impl<
                 coordinate = in(reg) &coordinate,
             }
         }
-        result
+        result.truncate_into()
     }
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, { Multisampled::False as u32 }, SAMPLED, FORMAT>
+        const COMPONENTS: u32,
+    >
+    Image<
+        SampledType,
+        DIM,
+        DEPTH,
+        ARRAYED,
+        { Multisampled::False as u32 },
+        SAMPLED,
+        FORMAT,
+        COMPONENTS,
+    >
 {
     // Note: #[inline] is needed because in vulkan, the component must be a constant expression.
     /// Gathers the requested component from four texels.
@@ -162,12 +190,12 @@ impl<
         sampler: Sampler,
         coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
         component: u32,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         Self: HasGather,
         F: Float,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
         unsafe {
             asm! {
                 "%typeSampledImage = OpTypeSampledImage typeof*{this}",
@@ -184,7 +212,7 @@ impl<
                 component = in(reg) component,
             }
         }
-        result
+        result.truncate_into()
     }
 
     /// Sample texels at `coord` from the image using `sampler`.
@@ -193,12 +221,12 @@ impl<
         &self,
         sampler: Sampler,
         coord: impl ImageCoordinate<F, DIM, ARRAYED>,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
         unsafe {
-            let mut result = Default::default();
+            let mut result = SampledType::Vec4::default();
             asm!(
                 "%typeSampledImage = OpTypeSampledImage typeof*{1}",
                 "%image = OpLoad typeof*{1} {1}",
@@ -212,7 +240,7 @@ impl<
                 in(reg) &sampler,
                 in(reg) &coord
             );
-            result
+            result.truncate_into()
         }
     }
 
@@ -224,12 +252,13 @@ impl<
         sampler: Sampler,
         coord: impl ImageCoordinate<F, DIM, ARRAYED>,
         bias: f32,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
         unsafe {
-            let mut result = Default::default();
+            let mut result = SampledType::Vec4::default();
+
             asm!(
                 "%typeSampledImage = OpTypeSampledImage typeof*{1}",
                 "%image = OpLoad typeof*{1} {1}",
@@ -244,7 +273,7 @@ impl<
                 in(reg) &coord,
                 in(reg) bias,
             );
-            result
+            result.truncate_into()
         }
     }
 
@@ -257,11 +286,11 @@ impl<
         sampler: Sampler,
         coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
         lod: f32,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
         unsafe {
             asm!(
                 "%image = OpLoad _ {this}",
@@ -278,7 +307,7 @@ impl<
                 lod = in(reg) &lod
             );
         }
-        result
+        result.truncate_into()
     }
 
     #[crate::macros::gpu_only]
@@ -290,11 +319,11 @@ impl<
         coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
         gradient_dx: impl ImageCoordinate<F, DIM, { Arrayed::False as u32 }>,
         gradient_dy: impl ImageCoordinate<F, DIM, { Arrayed::False as u32 }>,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
         unsafe {
             asm!(
                 "%image = OpLoad _ {this}",
@@ -313,7 +342,7 @@ impl<
                 gradient_dy = in(reg) &gradient_dy,
             );
         }
-        result
+        result.truncate_into()
     }
 
     #[crate::macros::gpu_only]
@@ -424,11 +453,12 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const SAMPLED: u32,
         const FORMAT: u32,
+        const COMPONENTS: u32,
     >
     Image<
         SampledType,
@@ -438,6 +468,7 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
     /// Sample the image with a project coordinate
@@ -447,12 +478,12 @@ impl<
         &self,
         sampler: Sampler,
         project_coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True as u32 }>,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
         unsafe {
-            let mut result = Default::default();
+            let mut result = SampledType::Vec4::default();
             asm!(
                 "%image = OpLoad _ {this}",
                 "%sampler = OpLoad _ {sampler}",
@@ -465,7 +496,7 @@ impl<
                 sampler = in(reg) &sampler,
                 project_coordinate = in(reg) &project_coordinate,
             );
-            result
+            result.truncate_into()
         }
     }
 
@@ -644,22 +675,36 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const FORMAT: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::No as u32 }, FORMAT>
+        const COMPONENTS: u32,
+    >
+    Image<
+        SampledType,
+        DIM,
+        DEPTH,
+        ARRAYED,
+        MULTISAMPLED,
+        { Sampled::No as u32 },
+        FORMAT,
+        COMPONENTS,
+    >
 {
     /// Read a texel from an image without a sampler.
     #[crate::macros::gpu_only]
     #[doc(alias = "OpImageRead")]
-    pub fn read<I>(&self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED>) -> SampledType::Vec4
+    pub fn read<I>(
+        &self,
+        coordinate: impl ImageCoordinate<I, DIM, ARRAYED>,
+    ) -> SampledType::SampleResult
     where
         I: Integer,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
 
         unsafe {
             asm! {
@@ -673,7 +718,7 @@ impl<
             }
         }
 
-        result
+        result.truncate_into()
     }
 
     /// Write a texel to an image without a sampler.
@@ -699,22 +744,36 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::Unknown as u32 }, FORMAT>
+        const COMPONENTS: u32,
+    >
+    Image<
+        SampledType,
+        DIM,
+        DEPTH,
+        ARRAYED,
+        MULTISAMPLED,
+        { Sampled::Unknown as u32 },
+        FORMAT,
+        COMPONENTS,
+    >
 {
     /// Read a texel from an image without a sampler.
     #[crate::macros::gpu_only]
     #[doc(alias = "OpImageRead")]
-    pub fn read<I>(&self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED>) -> SampledType::Vec4
+    pub fn read<I>(
+        &self,
+        coordinate: impl ImageCoordinate<I, DIM, ARRAYED>,
+    ) -> SampledType::SampleResult
     where
         I: Integer,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
 
         unsafe {
             asm! {
@@ -728,7 +787,7 @@ impl<
             }
         }
 
-        result
+        result.truncate_into()
     }
 
     /// Write a texel to an image without a sampler.
@@ -754,11 +813,12 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const FORMAT: u32,
+        const COMPONENTS: u32,
     >
     Image<
         SampledType,
@@ -768,6 +828,7 @@ impl<
         MULTISAMPLED,
         { Sampled::No as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
     /// Read a texel from subpass input attachment.
@@ -777,11 +838,11 @@ impl<
     pub fn read_subpass<I>(
         &self,
         coordinate: impl ImageCoordinateSubpassData<I, ARRAYED>,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         I: Integer,
     {
-        let mut result = Default::default();
+        let mut result = SampledType::Vec4::default();
 
         unsafe {
             asm! {
@@ -795,19 +856,20 @@ impl<
             }
         }
 
-        result
+        result.truncate_into()
     }
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
         const FORMAT: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, SAMPLED, FORMAT>
+        const COMPONENTS: u32,
+    > Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, SAMPLED, FORMAT, COMPONENTS>
 {
     /// Query the number of mipmap levels.
     #[crate::macros::gpu_only]
@@ -888,13 +950,24 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
         const FORMAT: u32,
-    > Image<SampledType, DIM, DEPTH, ARRAYED, { Multisampled::False as u32 }, SAMPLED, FORMAT>
+        const COMPONENTS: u32,
+    >
+    Image<
+        SampledType,
+        DIM,
+        DEPTH,
+        ARRAYED,
+        { Multisampled::False as u32 },
+        SAMPLED,
+        FORMAT,
+        COMPONENTS,
+    >
 {
     /// Query the dimensions of Image, with no level of detail.
     #[crate::macros::gpu_only]
@@ -922,11 +995,12 @@ impl<
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
         const FORMAT: u32,
+        const COMPONENTS: u32,
     >
     Image<
         SampledType,
@@ -936,6 +1010,7 @@ impl<
         { Multisampled::True as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
     /// Query the number of samples available per texel fetch in a multisample image.
@@ -967,70 +1042,76 @@ pub struct SampledImage<I> {
 }
 
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DIM: u32,
         const DEPTH: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
         const FORMAT: u32,
+        const COMPONENTS: u32,
     >
     SampledImage<
-        Image<SampledType, DIM, DEPTH, ARRAYED, { Multisampled::False as u32 }, SAMPLED, FORMAT>,
+        Image<
+            SampledType,
+            DIM,
+            DEPTH,
+            ARRAYED,
+            { Multisampled::False as u32 },
+            SAMPLED,
+            FORMAT,
+            COMPONENTS,
+        >,
     >
 {
     /// Sample texels at `coord` from the sampled image with an implicit lod.
-    ///
-    /// # Safety
-    /// Sampling with a type (`S`) that doesn't match the image's image format
-    /// will result in undefined behaviour.
     #[crate::macros::gpu_only]
-    pub unsafe fn sample<F>(
+    pub fn sample<F>(
         &self,
         coord: impl ImageCoordinate<F, DIM, ARRAYED>,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
-        let mut result = Default::default();
-        asm!(
-            "%sampledImage = OpLoad typeof*{1} {1}",
-            "%coord = OpLoad typeof*{2} {2}",
-            "%result = OpImageSampleImplicitLod typeof*{0} %sampledImage %coord",
-            "OpStore {0} %result",
-            in(reg) &mut result,
-            in(reg) self,
-            in(reg) &coord
-        );
-        result
+        let mut result = SampledType::Vec4::default();
+        unsafe {
+            asm!(
+                "%sampledImage = OpLoad typeof*{1} {1}",
+                "%coord = OpLoad typeof*{2} {2}",
+                "%result = OpImageSampleImplicitLod typeof*{0} %sampledImage %coord",
+                "OpStore {0} %result",
+                in(reg) &mut result,
+                in(reg) self,
+                in(reg) &coord
+            );
+        }
+        result.truncate_into()
     }
 
     /// Sample texels at `coord` from the sampled image with an explicit lod.
-    ///
-    /// # Safety
-    /// Sampling with a type (`S`) that doesn't match the image's image format
-    /// will result in undefined behaviour.
     #[crate::macros::gpu_only]
-    pub unsafe fn sample_by_lod<F>(
+    pub fn sample_by_lod<F>(
         &self,
         coord: impl ImageCoordinate<F, DIM, ARRAYED>,
         lod: f32,
-    ) -> SampledType::Vec4
+    ) -> SampledType::SampleResult
     where
         F: Float,
     {
-        let mut result = Default::default();
-        asm!(
-            "%sampledImage = OpLoad typeof*{1} {1}",
-            "%coord = OpLoad typeof*{2} {2}",
-            "%lod = OpLoad typeof*{3} {3}",
-            "%result = OpImageSampleExplicitLod typeof*{0} %sampledImage %coord Lod %lod",
-            "OpStore {0} %result",
-            in(reg) &mut result,
-            in(reg) self,
-            in(reg) &coord,
-            in(reg) &lod,
-        );
-        result
+        let mut result = SampledType::Vec4::default();
+        unsafe {
+            asm!(
+                "%sampledImage = OpLoad typeof*{1} {1}",
+                "%coord = OpLoad typeof*{2} {2}",
+                "%lod = OpLoad typeof*{3} {3}",
+                "%result = OpImageSampleExplicitLod typeof*{0} %sampledImage %coord Lod %lod",
+                "OpStore {0} %result",
+                in(reg) &mut result,
+                in(reg) self,
+                in(reg) &coord,
+                in(reg) &lod,
+            );
+        }
+        result.truncate_into()
     }
 }
 
@@ -1041,11 +1122,12 @@ impl<
 /// `OpTypeImage` must be 0."
 pub trait HasGather {}
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasGather
     for Image<
         SampledType,
@@ -1055,15 +1137,17 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasGather
     for Image<
         SampledType,
@@ -1073,15 +1157,17 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasGather
     for Image<
         SampledType,
@@ -1091,6 +1177,7 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
@@ -1101,12 +1188,13 @@ impl<
 /// "Its Dim operand must be one of 1D, 2D, 3D, or Cube."
 pub trait HasQueryLevels {}
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQueryLevels
     for Image<
         SampledType,
@@ -1116,16 +1204,18 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQueryLevels
     for Image<
         SampledType,
@@ -1135,16 +1225,18 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQueryLevels
     for Image<
         SampledType,
@@ -1154,16 +1246,18 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQueryLevels
     for Image<
         SampledType,
@@ -1173,6 +1267,7 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
@@ -1184,11 +1279,12 @@ impl<
 /// 3D, or Cube, it must also have either an MS of 1 or a Sampled of 0 or 2."
 pub trait HasQuerySize {}
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1198,11 +1294,17 @@ impl<
         { Multisampled::True as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::OneD as u32 },
@@ -1211,11 +1313,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::Unknown as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::OneD as u32 },
@@ -1224,15 +1332,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::No as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1242,11 +1352,17 @@ impl<
         { Multisampled::True as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::TwoD as u32 },
@@ -1255,11 +1371,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::Unknown as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::TwoD as u32 },
@@ -1268,15 +1390,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::No as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1286,11 +1410,17 @@ impl<
         { Multisampled::True as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::ThreeD as u32 },
@@ -1299,11 +1429,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::Unknown as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::ThreeD as u32 },
@@ -1312,15 +1448,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::No as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1330,11 +1468,17 @@ impl<
         { Multisampled::True as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::Cube as u32 },
@@ -1343,11 +1487,17 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::Unknown as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
-impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32>
-    HasQuerySize
+impl<
+        SampledType: SampleType<FORMAT, COMPONENTS>,
+        const DEPTH: u32,
+        const FORMAT: u32,
+        const ARRAYED: u32,
+        const COMPONENTS: u32,
+    > HasQuerySize
     for Image<
         SampledType,
         { Dimensionality::Cube as u32 },
@@ -1356,16 +1506,18 @@ impl<SampledType: SampleType<FORMAT>, const DEPTH: u32, const FORMAT: u32, const
         { Multisampled::False as u32 },
         { Sampled::No as u32 },
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1375,16 +1527,18 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const MULTISAMPLED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySize
     for Image<
         SampledType,
@@ -1394,6 +1548,7 @@ impl<
         MULTISAMPLED,
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
@@ -1404,11 +1559,12 @@ impl<
 /// "Its Dim operand must be one of 1D, 2D, 3D, or Cube, and its MS must be 0."
 pub trait HasQuerySizeLod {}
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySizeLod
     for Image<
         SampledType,
@@ -1418,15 +1574,17 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySizeLod
     for Image<
         SampledType,
@@ -1436,15 +1594,17 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySizeLod
     for Image<
         SampledType,
@@ -1454,15 +1614,17 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }
 impl<
-        SampledType: SampleType<FORMAT>,
+        SampledType: SampleType<FORMAT, COMPONENTS>,
         const DEPTH: u32,
         const FORMAT: u32,
         const ARRAYED: u32,
         const SAMPLED: u32,
+        const COMPONENTS: u32,
     > HasQuerySizeLod
     for Image<
         SampledType,
@@ -1472,6 +1634,7 @@ impl<
         { Multisampled::False as u32 },
         SAMPLED,
         FORMAT,
+        COMPONENTS,
     >
 {
 }